Código ordenado, bugs corregidos.

+++ MemeFinder/src/modelo/ModeloPrincipal.java
@@ -2,19 +2,14 @@ package modelo;
 
 import java.io.File;
 import java.nio.file.Files;
-import java.nio.file.Path;
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.List;
 
 import org.bson.Document;
 import org.bson.conversions.Bson;
 import org.bson.types.ObjectId;
 
 import com.mongodb.BasicDBObject;
-import com.mongodb.Cursor;
-import com.mongodb.DBCursor;
 import com.mongodb.MongoClient;
 import com.mongodb.client.FindIterable;
 import com.mongodb.client.MongoCollection;
@@ -63,15 +58,6 @@ public class ModeloPrincipal {
 			mdb = mc.getDatabase("MemeFinder3019");
 			tablaImagenes = mdb.getCollection("tablaImagenesTEST");
 			tablaEtiquetas = mdb.getCollection("tablaEtiquetasTEST");
-			//System.out.println(mdb.getName());
-			//System.out.println(mdb.getCollection("tablaImagenesTEST").countDocuments());
-			//mc.close();
-			
-//			DBCursor cur = tablaImagenes.find();
-			
-			//Document document = new Document();
-			
-			
 			
 		}catch(Exception e) {
 			e.printStackTrace();
@@ -84,15 +70,6 @@ public class ModeloPrincipal {
 	 public boolean anadirImagen(ImagenTemp imagenTemp) {
 		 
 		 //Modifica tablaImagenes
-		 
-//		 File temp = null;
-//			do {
-//				temp = new File(rutaTemp + Math.round((Math.random()*1000)) + archivosElegidos[i].getName().substring(archivosElegidos[i].getName().lastIndexOf('.')));
-//				
-//			}while(Files.exists(temp.toPath()));
-//			
-//			try {
-//				Files.copy(archivosElegidos[i].toPath(), temp.toPath());
 		 
 		 File temp = null;
 		 String nombreExt = "";
@@ -124,34 +101,7 @@ public class ModeloPrincipal {
 		 //TODO comprobar que no falla si getArrayEtiquetas().isEmpty()
 		 for(Etiquetas e : imagenTemp.getArrayEtiquetas()) {
 			 
-			 anadirEtiqueta(e.getTexto());
-			 
-//			 //Si no existe esa etiqueta en BD
-//			 if(tablaEtiquetas.find(new Document(_ID,e.getTexto())).first() == null) {
-////				 System.out.println("etiquetas null");
-//			 //if(tablaEtiquetas.countDocuments(new Document(_ID,e.getTexto())) == 0) {
-//				 
-//				 Document doc = new Document();
-//				 doc.append(_ID, e.getTexto());
-//				 doc.append(ETCOUNT, 1);
-//				 
-//				 tablaEtiquetas.insertOne(doc);
-//				 
-////				 System.out.println("count0: " + tablaEtiquetas.find(new Document(_ID,e.getTexto())).first());
-//				 
-//			 }else {
-//				 System.out.println("etiquetas no null");
-//				 Document newDocument = new Document();
-//				 newDocument.append("$inc", new BasicDBObject().append(ETCOUNT, 1));
-//								
-//						tablaEtiquetas.updateOne(new Document().append(_ID, e.getTexto()), newDocument);
-//						
-////						System.out.println("count+: " + tablaEtiquetas.find(new Document(_ID,e.getTexto())).first());
-//				 
-//			 }
-			 
-			 
-			 
+			 anadirEtiqueta(e.getTexto()); 
 			 
 		 }
 		 
@@ -166,12 +116,7 @@ public class ModeloPrincipal {
 		 
 		 //ojo !
 		 if(!nuevoNombre.isEmpty()) {
-			 System.out.println("entraaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa");
-			 
 			 UpdateResult aaa = tablaImagenes.updateOne(eq(_ID,tempId), Updates.set(NOMBRE, (nuevoNombre + it.getExtension())));
-			 System.out.println("matches: " + aaa.getMatchedCount());
-			 System.out.println("modificados: " + aaa.getModifiedCount());
-			 
 		 }
 		 
 		//ojo !
@@ -192,41 +137,6 @@ public class ModeloPrincipal {
 		 
 	 }
 	 
-//	 public void updateImagen(ImagenTemp it, String nuevoNombre, ArrayList<Etiquetas> nuevoArrayEtiquetas) {
-//		 
-//		 //obsoleto?
-//		 
-//		 //ojo !
-//		 if(!it.getNombre().equals(nuevoNombre)) {
-//			 tablaImagenes.updateOne(eq(_ID,it.getIdMongo()), Updates.set(NOMBRE, nuevoNombre));
-//		 }
-//		 
-//		 //bool[] con length del array etiquetas antiguo, será false solo en las etiquetas eliminadas de la imagen
-//		 boolean[] arrayIndicesEtBorradas = new boolean[it.getArrayEtiquetas().size()];
-//		 for(int i=0; i<arrayIndicesEtBorradas.length; i++) arrayIndicesEtBorradas[i]=false;
-//		 
-//		 for(Etiquetas et : nuevoArrayEtiquetas) {
-//			 
-//			 //Si arrayviejo contiene la etiqueta et (de arraynuevo), se marca true en el bool[]
-//			 if(it.getArrayEtiquetas().contains(et)) {
-//				 arrayIndicesEtBorradas[it.getArrayEtiquetas().indexOf(et)] = true;
-//			
-//			 //Si arraynuevo contiene una et que arrayviejo no, se añade la etiqueta	 
-//			 }else {
-//				 anadirEtiqueta(et.getTexto());
-//			 }
-//		 }
-//		 
-//		 //Se recorre el bool[], y en los casos false, se borra la etiqueta, porque serán
-//		 //los únicos en los que no han sido cambiados a true por el foreach anterior
-//		 for(int i=0; i<it.getArrayEtiquetas().size(); i++) {
-//			 if(arrayIndicesEtBorradas[i] == false) {
-//				 borrarEtiqueta(it.getArrayEtiquetas().get(i).getTexto());
-//			 }
-//		 }
-//		 
-//	 }
-	 
 	 public boolean borrarImagen(ObjectId id) {
 		 
 		 try {
@@ -236,12 +146,6 @@ public class ModeloPrincipal {
 			 for(String s : (ArrayList<String>) doc.get(ETIQUETAS)) {
 				 
 				 borrarEtiqueta(s);
-				 
-//				 if(getCountEtiquetas(s) == 1) {
-//					 tablaEtiquetas.deleteOne(eq(_ID,s));
-//				 }else {
-//					 tablaEtiquetas.updateOne(eq(_ID,s),Updates.inc(ETCOUNT, -1));
-//				 }
 	 
 			 }
 			 
@@ -261,8 +165,6 @@ public class ModeloPrincipal {
 		 
 		 //Si no existe esa etiqueta en BD
 		 if(tablaEtiquetas.find(new Document(_ID,strEt)).first() == null) {
-//			 System.out.println("etiquetas null");
-		 //if(tablaEtiquetas.countDocuments(new Document(_ID,e.getTexto())) == 0) {
 			 
 			 Document doc = new Document();
 			 doc.append(_ID, strEt);
@@ -270,16 +172,14 @@ public class ModeloPrincipal {
 			 
 			 tablaEtiquetas.insertOne(doc);
 			 
-//			 System.out.println("count0: " + tablaEtiquetas.find(new Document(_ID,e.getTexto())).first());
 			 
 		 }else {
 			 System.out.println("etiquetas no null");
 			 Document newDocument = new Document();
 			 newDocument.append("$inc", new BasicDBObject().append(ETCOUNT, 1));
 							
-					tablaEtiquetas.updateOne(new Document().append(_ID, strEt), newDocument);
-					
-//					System.out.println("count+: " + tablaEtiquetas.find(new Document(_ID,e.getTexto())).first());
+			 tablaEtiquetas.updateOne(new Document().append(_ID, strEt), newDocument);
+//			 System.out.println("count+: " + tablaEtiquetas.find(new Document(_ID,e.getTexto())).first());
 			 
 		 }
 		 
@@ -288,11 +188,8 @@ public class ModeloPrincipal {
 	 //No borra per se, resta 1 de la tabla tablaEtiquetas y borra si getCount==1
 	 private void borrarEtiqueta(String strEt) {
 		 
-		 if(getCountEtiquetas(strEt) == 1) {
-		 tablaEtiquetas.deleteOne(eq(_ID,strEt));
-	 }else {
-		 tablaEtiquetas.updateOne(eq(_ID,strEt),Updates.inc(ETCOUNT, -1));
-	 }
+		 if(getCountEtiquetas(strEt) == 1) tablaEtiquetas.deleteOne(eq(_ID,strEt));
+		 else tablaEtiquetas.updateOne(eq(_ID,strEt),Updates.inc(ETCOUNT, -1));
 		 
 	 }
 	 
@@ -300,13 +197,8 @@ public class ModeloPrincipal {
 		 
 		 Document retorna = (Document) tablaEtiquetas.find(new Document(_ID,strEt)).first();
 		 
-		 if(retorna==null) {
-			 return 0;
-		 }else {
-			 return retorna.getInteger(ETCOUNT, 90);
-			 
-		 }
-		 
+		 if(retorna==null) return 0;
+		 else return retorna.getInteger(ETCOUNT, 90);
 		 
 	 }
 	 
@@ -314,18 +206,10 @@ public class ModeloPrincipal {
 	 public ArrayList<ImagenTemp> getArrayComparaciones(ImagenTemp it, int minPorcentaje) {
 		 
 		 ArrayList<ImagenTemp> arrayRetorna = new ArrayList<ImagenTemp>();
-		 //String hashOriginal = it.getPHash();
 		 Boolean[] hashOriginal = it.getPHash();
-		 
-//		 System.out.println();
-//		 System.out.println("hash original:  ");
-//		 for(int i=0; i<hashOriginal.length; i++) {
-//			 System.out.print(hashOriginal[i] + "-");
-//		 }
-//		 System.out.println();
+
 		 
 		 FindIterable<Document> iterable = tablaImagenes.find();
-//		 System.out.println("iterable: " + iterable.toString());
 		 FindIterable<BasicDBObject> caca = tablaImagenes.find();
 		 
 		 for(Document doc : iterable) {
@@ -338,15 +222,7 @@ public class ModeloPrincipal {
 				 
 			 }
 			 
-//			 System.out.println("comprado con: " + ((String)doc.get(NOMBRE)));
-			 //int compara = ProxyHash.getDistancia((Boolean[])(((ArrayList<Boolean>)doc.get(PHASH)).toArray()), hashOriginal);
-//			 int compara = ProxyHash.getDistancia(
-//					 (Boolean[])(doc.getList(PHASH,
-//					 Boolean.class).toArray()), hashOriginal);
-			 //investiga(doc.get(PHASH));
-			 float compara = ProxyHash.getDistancia(hashMeCagoEnLaHostia,
-					 hashOriginal);
-//			 System.out.println("compara: " + compara);
+			 float compara = ProxyHash.getDistancia(hashMeCagoEnLaHostia,hashOriginal);
 			 if(compara > (((float)minPorcentaje)/((float)100))) {
 				 ImagenTemp temp = convierteDocObjeto(doc);
 				 temp.setIndiceParecido(compara);
@@ -442,25 +318,9 @@ public class ModeloPrincipal {
 	 private ImagenTemp convierteDocObjeto(Document d) {
 		 
 		 String url = (String) d.get(URL);
-//		 System.out.println("la url es " + url);
 		 String nombreFull = (String) d.get(NOMBRE);
-//		 System.out.println("nombrefull: " + nombreFull);
-//		 //ArrayList<Etiquetas> etiquetas = (ArrayList<Etiquetas>) d.get(ETIQUETAS);
-//		 ArrayList<Etiquetas> etiquetas = new ArrayList<Etiquetas>();
-//		 for(String s : (String[]) d.get(ETIQUETAS)) {
-//			 etiquetas.add(new Etiquetas(s,getCountEtiquetas(s)));
-//		 }
-		 //String pHash = (String) d.get(PHASH);
+
 		 
-//		 Boolean[] pHash = (Boolean[])d.get(PHASH);
-		 
-//		 ArrayList<Boolean> arb = (ArrayList<Boolean>)doc.get(PHASH);
-//		 Boolean[] hashMeCagoEnLaHostia = new Boolean[arb.size()];
-//		 for(int i=0; i<arb.size(); i++) {
-//			 
-//			 hashMeCagoEnLaHostia[i] = arb.get(i);
-//			 
-//		 }
 		 ArrayList<Boolean> boolTemporal = (ArrayList<Boolean>) d.get(PHASH);
 		 Boolean[] pHash = new Boolean[boolTemporal.size()];
 		 for(int i=0; i<boolTemporal.size(); i++) pHash[i] = boolTemporal.get(i);
